backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
websockets
xmltodict
jinja2

# Dev tools
black
isort
pre-commit
